transform services to repository and add some creation functions
